review: move response creation function to update worker file
