+++ bin/proving-service/src/proxy/health_check.rs
@@ -1,11 +1,16 @@
+use std::time::Duration;
+
 use axum::async_trait;
 use pingora::{prelude::sleep, server::ShutdownWatch, services::background::BackgroundService};
+use tonic::transport::Channel;
+use tonic_health::pb::health_client::HealthClient;
 use tracing::debug_span;
 
 use super::{
     metrics::{WORKER_COUNT, WORKER_UNHEALTHY},
     LoadBalancerState,
 };
+use crate::error::TxProverServiceError;
 
 /// Implement the BackgroundService trait for the LoadBalancer
 ///
@@ -54,4 +59,28 @@ impl BackgroundService for LoadBalancerState {
         })
         .await;
     }
+}
+
+// HELPERS
+// ================================================================================================
+
+/// Create a gRPC [HealthClient] for the given worker address.
+///
+/// # Errors
+/// - [TxProverServiceError::InvalidURI] if the worker address is invalid.
+/// - [TxProverServiceError::ConnectionFailed] if the connection to the worker fails.
+pub async fn create_health_check_client(
+    address: String,
+    connection_timeout: Duration,
+    total_timeout: Duration,
+) -> Result<HealthClient<Channel>, TxProverServiceError> {
+    let channel = Channel::from_shared(format!("http://{}", address))
+        .map_err(|err| TxProverServiceError::InvalidURI(err, address.clone()))?
+        .connect_timeout(connection_timeout)
+        .timeout(total_timeout)
+        .connect()
+        .await
+        .map_err(|err| TxProverServiceError::ConnectionFailed(err, address))?;
+
+    Ok(HealthClient::new(channel))
 }
